[stable-only] Use bionic for requirements-check job
requirements-check jobs started to fail in stable/stein with

  ModuleNotFoundError: No module named 'openstack_requirements'

This is due to that job runs in ubuntu-focal instead of ubuntu-bionic,
and openstack_requirements is not installed for python 3.8. By pinning
the nodeset to bionic we can eliminate this issue as in bionic the
python version is 3.6.

NOTE(elod.illes): older stable branches fail, too. In rocky and older
branches Xenial was the supported nodeset, however the check job will
pass with Bionic, no need to use the older distro.

Change-Id: I077e6276618a257b2c948e58d4de993998844827
(cherry picked from commit 3a3ed8a7d36c2321a519fd6b54b7b328274e4d07)