Adding ansible-core to the blacklist
Ansible core is going to take the role of ansible in some projects,
as the dependencies move to the ansible>2.10.

Right now that's the case only for VF, but more might come in
the future and it's best to be ready.

The inclusion of the ansible-core on the blacklist would serve
exactly the same purpose as the inclusion of original ansible package.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I36331013568a36c963c180db10e30c3b72cf86a8