Add cross-placement jobs
A recent requirements bump of oslo.db from 10.0.0 => 11.0.0 [1] caused
failures in the placement project gate, in the functional tests.

This raised awareness that we have not had cross jobs checking with
placement since placement was split out from nova during the Train
cycle.

This adds cross-placement jobs which correspond to the existing
cross-nova ones.

Story: 2009209
Task: 43275

[1] I03d79b19774650f3df7129157e6a12998fd33447

Change-Id: I625b2345248f2198ea264e7cafbe6e93e4ba4a1e