Drop unbuildable package from global-requirements
suds-jurko is no longer buildable on a modern python3 environment, the only
reason that this didn't show up as gate failures so far is that we are
running on the opendev CI with a set of cached wheel builds, where
versions of this package were built with an older setup some time age
when things still worked. This will no longer be the case when switching
to python3.10.

See [0] for a test build without the wheel cache in use.

[0] https://review.opendev.org/813292

Change-Id: Ic9b4200ad98879c9d2bd4c0544ba9f1f3fb412f9