Add prettytable to global-requirements
PrettyTable package is an old naming of prettytable.
Currently prettytable and PrettyTable resolve to the same package
in pypi and both names work.
As a result, when installed it's shown as prettytable.

This patch aims to allow developers to add prettytable package to
their project requirements and not have to reffer to the old naming.

Needed-By: https://review.opendev.org/c/openstack/openstack-ansible/+/777061
Change-Id: I75a025d3c347cc895328949281b8068b7af85a6c