Add task-core to global-requirements.txt
The TripleO team discussed and agreed to move forward with adding task-core to
TripleO at the last PTG:
https://etherpad.opendev.org/p/tripleo-directord-task-core

* Is the library actively maintained?

Yes, it is actively maintained by the TripleO community at
https://github.com/directord/task-core

* Is the library good code?

Yes, as the authors, we'd like to think so.

* Is the library python 3 compatible?

Yes. There are python3 jobs setup on Github.

* Is the library license compatible?

Yes. It is licensed Apache 2.0.

* Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

Not yet, but we are in the process of packaging it in RDO:
https://bugzilla.redhat.com/show_bug.cgi?id=2021983

* Is the function of this library already covered by other libraries in global-requirements.txt?

No.

* Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?

Yes. TripleO.
The first patch that would require task-core is:
https://review.opendev.org/c/openstack/python-tripleoclient/+/821624

If the library release is managed by the Openstack release process does it use the cycle-with-intermediary release type?

It is not yet managed by the OpenStack release process, but may be in the future.

Do I need to update anything else?

No.

Signed-off-by: James Slagle <jslagle@redhat.com>
Change-Id: Ia7b92d62abdedcb6222cef4159d9edecdf65e0e0