Pin gabbi to <2.5.0
The recent change[1] in gabbi revealed an issue with urllib3 and now
http requests no longer works. This causes failure in some components
like heat-tempest-plugin[2].

This change pins version of gabbi to exclude that change until
the issue is fixed in urllib3[3].

Note that gabbi 2.5.0 has been yanked because of this known issue.

[1] https://github.com/cdent/gabbi/pull/306
[2] http://lists.openstack.org/pipermail/openstack-discuss/2022-January/026724.html
[3] https://github.com/urllib3/urllib3/issues/2534

Change-Id: Iaaf5d1615d57ae9e53a0c64a0bcf0aba02144bf0