Bump pymemcache to 3.5.0
Bumping the version to get new feature and fixes [1][2][3][4] need
by oslo.cache [5][6] to fix socket deconnection recover of a
memcached cluster node.

I know that we usually avoid to bump requirements on stable branches
but without the changes landed by dogpile.cache 1.1.5 and
pymemcache 3.5.0 oslo.cache won't be able to recover from a connection
issue in a Memcached cluster. This a problem for an HA context.

In other words if the used node of the cluster become down, oslo.cache
won't switch to another node and won't recover. That will lead us to
this bug https://bugs.launchpad.net/oslo.cache/+bug/1959562

It will have for side for openstack to break TLS support with
memcached.  Indeed we introduced the pymemcache in oslo.cache to
allow support TLS. This backend was precedently backported up
to train.

The changes landed by pymemcache 3.5.0 are new isolated (optional)
features so I won't expect side effects introduced by these changes.
The goal of the pymemcache changes is to allow us to configure
manually the retry mechanisms and socket keepalives, that's all.

Also a bug fix is landed by pymemcache 3.5.0. This fix close
connection properly on errors.

[1] https://github.com/sqlalchemy/dogpile.cache/commit/da490df9edae8c60e7ee3d577ac610efdffc5986

[1] https://github.com/pinterest/pymemcache/commit/07b5ecc21ce5d388d4312c943d79f813311e349f
[2] https://github.com/pinterest/pymemcache/commit/75fe5c81c35d2bcfc8e6a697aef948efbfebe8ba
[3] https://github.com/pinterest/pymemcache/commit/15c35ef184c09279cbb886ceb4bd2807241d48b1
[4] https://github.com/pinterest/pymemcache/commit/b289c87bb89b3ab477bd5d92c8951ab42c923923
[5] https://review.opendev.org/c/openstack/oslo.cache/+/803747
[6] https://review.opendev.org/c/openstack/oslo.cache/+/803716

Partial-Bug: #1959562
Change-Id: I6859a91125c94f548213cc222f3e495ed6191e85