Revert "Block neutron-lib 2.21.0"
This reverts commit 5a8e4c5122389d010b74d0d8c84336b3ceb142ed.

Reason for revert:
The error found in LP#1973162 and LP#1942190 seems to be in the Neutron
L3 code. There is a race condition when two routers try to update a HA
port. With [1][2], this is more evident as we protect the DB transactions
and Neutron always creates two active port bindings for the same port.

[3] is addressing this problem. I'm reverting this patch to test [3]
with neutron-lib 2.21.0

[1]https://review.opendev.org/c/openstack/neutron-lib/+/828738
[2]https://review.opendev.org/c/openstack/neutron/+/833247
[3]https://review.opendev.org/c/openstack/neutron/+/841712

Change-Id: I099703bb72291afa9cb7a5130b823334303528d5