Revert "Add prettytable to global-requirements"
This reverts commit 05996e009c55f757cf91cf0d94abd86dcb0f21c4.

Reason for revert:

From what I can tell, instead of allowing you to use either PrettyTable or prettytable in your requirements.txt, it forces you to use *both*, which is completely contrary to the intention of this patch.  See the commit message on https://review.opendev.org/c/openstack/glance/+/843143 for how I came to this conclusion.

The patch that the original change was Needed-By never merged (it failed the requirements-check job).

Being required to use both PrettyTable *and* prettytable in requirements.txt is probably a bug in the requirements checking code, but that shouldn't stop us from reverting this patch.

Change-Id: Id65e566b8f943cf60d4f2b646e88876c4ababbd4