[stable-only]Pin PrettyTable in py3.9 in stable/yoga
Since the latest PrettyTable 3.4.0 release the
openstack-tox-py39 job is failing constantly on stable/yoga in
nova. It seems that there is a breaking change in the latest
PrettyTable release. As this is affecting a stable branch this patch
proposes to pin the dependency to the last known good version on the
affected branch.

Closes-Bug: #1988482
Change-Id: I757b6f9f5bbfbca214b962b00a16f8f9c06657fa