Cap virtualenv (<20.18.0) to unblock gate
With latest virtualenv (and setuptools bundled in it) py3*-check-uc
jobs started to fail as pkg_resources' __getattr__ breaks on multiple
packages' metadata.

Release is very close and library release upper constraints bumping
patches are piling up. This patch is a workaround only to unblock the
gate and need to be reverted as soon as the py3*-check-uc jobs works
well with the latest virtualenv and setuptools releases.

Change-Id: I36ae32a401844bf5a5df0d20d3f88f24b2a0cf8e