Fix gate by pinning hacking
Latest 'hacking' releases (version>=6.0) introduced newer flake8 that
started to cause fail for pep8 tox target. For this time it could be
easily fixed but since this is an old stable branch which is already
in Extended Maintenance, the more stable solution is to pin 'hacking'
version to the last that is still makes pep8 passing.

Changes:
  .zuul.d/project.yaml

NOTE(elod.illes): openstacksdk-functional-devstack job is failing
because it overrides openstacksdk's branch to master branch, in which
the dependency constraints contradicts with the constraints on wallaby
branch, thus we need to override here the branch back to the actual
stable branch.

Change-Id: I47662329e89899dbfa145478bc23c61da4c031f3
(cherry picked from commit c2d7b687aca78602ac16f5276806b3b5b398f21d)