Fix gate by pinning hacking
Latest 'hacking' releases (version>=6.0) introduced newer flake8 that
started to cause fail for pep8 tox target. For this time it could be
easily fixed but since this is an old stable branch which is already
in Extended Maintenance, the more stable solution is to pin 'hacking'
version to the last that is still makes pep8 passing.

Change-Id: I47662329e89899dbfa145478bc23c61da4c031f3