Pin XStatic-Font-Awesome to <4.8.0.0
This patch limits the version of XStatic-Font-Awesome in
``global-requirements.txt`` to <4.8.0. The purpose is to
prevent automatic updates of XStatic-Font-Awesome, as the
Horizon team does not currently support or require the
updated versions. This decision is based on the fact that
XStatic-Font-Awesome, being a font package[1], does not pose
any security risks or vulnerabilities (CVEs). By maintaining
the current version, we ensure compatibility and stability
within the Horizon project.

[1] https://meetings.opendev.org/irclogs/%23openstack-horizon/%23openstack-horizon.2023-05-31.log.html#t2023-05-31T15:50:28

Change-Id: Iadc94128566ee68da13eb6fd4f9dd20e2e3d4b35