Add platformdirs to global requirements
In preparation to migrate projects from the abandoned appdirs
library to the currently maintained platformdirs replacement,
explicitly add it to our global requirements list. It's already in
the upper constraints list due to being a transitive dependency, but
we want to make sure we don't inadvertently drop it in the future.

Change-Id: I58c5c36f2f60030f81b09fe27b68323f7f46caeb