Raise cap for XStatic-JQuery in ``global-requirements.txt``
This patch raise the version of XStatic-JQuery <3.6
in ``global-requirements.txt`` file because horizon
is already using XStatic-JQuery-Migrate version 3.3.2.1 [1]
but XStatic-jQuery old version is not compatiable with
latest version of XStatic-JQuery-Migrate version.
So horizon is updating XStatic-JQuery to <3.6 and for that
this patch needs to be merge first.

Note: ``upper-constraints.txt`` is updated in a seprate patch
[3].

[1] https://review.opendev.org/c/openstack/requirements/+/883402
[2] https://review.opendev.org/c/openstack/horizon/+/887548
[3] https://review.opendev.org/c/openstack/requirements/+/887933

Change-Id: Iccd7166f0a8091bdf4613186f9c8048864603f29