Increase upper-constraint for PBR
pbr in stable/victoria uses functions that were deprecated and
removed from setuptools > 68.0.0

This is causing some stable/victoria gates to fail because pbr is unable
to find the functions that were removed.

The 5.5.1 release includes fixes that should be able to handle the
removal of the deprecated functions in setuptools.

Change-Id: I23a4333ed8c2ede5cfe2d392abcb3722d4d0ae81