Allow locally requiring a package for specific python versions
Sometimes we have to require a specific library in certain python
versions only.

In oslo and a few other projects, we are now replacing pytz by built-in
zoneinfo[1] to reduce 3rd party dependency, but zoneinfo is not
available in python < 3.9 and pytz is still required in python < 3.9.
In addition we need tzdata when zoneinfo is used, which means we need
tzdata only in python >= 3.9.

However these updates are currently prohibited by requirement check
because pytz in the global requirement does not contain python version
marker.

This change allows having a local requirement with python marker along
with the corresponding global requirement without python marker to
support such usage.

[1] https://review.opendev.org/c/openstack/oslo.utils/+/897623

Change-Id: I2bd5671962b3ca1abac91c2afba6e2c566616390