Downgrade dnspython from 2.6.0 to 2.5.0
As stated in an upstream bug report[1]:

"Dnspython 2.6.0 has been released. This may cause software using the
latest dnspython and eventlet to fail with timeouts, as the monkey
patching in eventlet has not yet been updated. The eventlet team is
aware of the dnspython release and a patch is in progress. In the
meantime, affected users should pin dnspython to < 2.6."

[1] https://github.com/rthalley/dnspython/issues/1050

Change-Id: I993dc7b71f2629d04a10229e429f613bde134965