Remove Xenial and CentOS 7 wheel cache build jobs
The OpenDev team is looking towards removing CentOS 7 and Ubuntu Xenial
test nodes from Zuul and Nodepool. In the case of Xenial this distro
release has been EOL for some time now, and in the case of CentOS 7 will
go EOL in the near future. Removing these distro releases will allow the
OpenDev team to focus on building images for modern releases as well as
freeing up space on our mirror filesystems for additional mirror
content.

At this point OpenStack should largely be on CentOS Stream and Ubuntu
Bionic, Focal, and Jammy. We don't expect this to create a major impact.

Change-Id: I1ca9403525df470a7da8c1f02bdfb8d4ea90c670