Add some more types packages
The added packages are used in [1] to allow mypy do proper type checking
in the Octavia project. The specific versions in u-c is needed in
order to avoid false positives due to unaligned version of the
respective types-* packages.

This is somewhat related to the discussion on the mailing list about
modernization [2] even though I started working on it in Octavia way
before that discussion started.

types-requests===2.31.0.6 is necessary because we're still using
urllib3<2.[3]

[1]: https://review.opendev.org/c/openstack/octavia/+/879749
[2]: https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/thread/4V63CHMZ4GPC4IYN7JCJPVKHLZAHN5BL/
[3]: https://pypi.org/project/types-requests/#description

Change-Id: Ic8b557a512c29ae76a3ba5e805f7b6f5d118baff