Add python-consul2 to global requirements
python-consul2 was added to tooz some time ago to support Consul
backend. Unfortunately it was not added to the global requirements at
that time and now prevents us from adding the requirement-check job
to tooz. This proposes adding the library to global requirements so
that we can add requirements check now to avoid further unmanaged
requirements. Note that the library has been inactive for a few years
and we are now discussing deprecation of the backend.

- Is the library actively maintained?
No. We are discussing our plan to remove/replace it.

- Is the library good code?
Yes

- Is the library python 3 compatible?
Yes

- Is the library license compatible?
Yes, MIT.

- Is the library already packaged in the distros we target
  (Ubuntu latest / Fedora latest)?
Not yet.

- Is the function of this library already covered by other libraries
  in global-requirements.txt?
No

- Is the library required for OpenStack project or related dev
  or infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes. It's required by tooz.

- If the library release is managed by the Openstack release process
  does it use the cycle-with-intermediary release type?
No

Change-Id: I44eb8f052ecc9872b22f0eb815351c2925e01bfa