Update cross-horizon-npm job
Horizon has dropped support for python3.8, so we need to run this job on
a node that has a more recent python version. This also needs to use a
more recent node version, then. Make the job inherit the same parent job
definition that horizon is using in order to avoid duplication. Drop our
generic requirements-cross-npm template since it is no longer being
used.

Change-Id: I0ea11984e63b885668ed5842e1c6221739e18343