Re-add python3.9 to upper-constraints
This patch partially reverts https://review.opendev.org/c/openstack/requirements/+/948285
re-adding python3.9 compatible libraries to upper-constraints.
This is done to give time to projects that still run and support
Python3.9 (for example any container/CI jobs based on CentOS Stream 9)
to phase off its usage and pass to a more recent Python version.

Python 3.9 is expect to EoL in October, we expect to be able to
remove its support once and for all before then.

Change-Id: Ie03307aac1dffc5cf397a2513c2c44d403a2d2b7