Drop build-wheel-cache-ubuntu-bionic-arm64
The OpenDev team is dropping bionic arm64 image builds as part of the
nodepool in zuul migration. It seems that much of the use of this
platform on arm64 was for nonvoting purposes and wasn't really used for
anything important. As part of this cleanup we need to remove the wheel
cache building for the platform which necessitates this cleanup.

Change-Id: Ic14bfa262eb0a63d9cd2eac719c73b7681169975
Signed-off-by: Clark Boylan <clark.boylan@gmail.com>