Pin boto3 and botocore to 1.35.99
Starting with 1.36.0[1], boto3 implements a new security feature that
may not be implemented by some S3-compatible storage backends, such as
Ceph RGW. This causes glance_store users to notice failures when
uploading images to such backends[2]. This might also impact other projects
such as Cinder.

Pinning boto3 to 1.35.99 until we come up with a proper workaround for
this. Also pinning botocore to 1.35.99 as this is the only version that
is compatible with boto3 1.35.99, and pinning s3transfer to 0.10.0 as
this is the only version that is compatible with botocore 1.35.99.

(tonyb) As this is potentially a longer term issue, include the caps
in global-requirements.txt to ease constraints updates.

[1] https: //github.com/boto/boto3/blob/dd74a1e97c450cf7670cabf3a9516b5127bcffb5/CHANGELOG.rst#L1882
[2] https://bugs.launchpad.net/glance/+bug/2121144

Change-Id: I311a74d6839410779f89f3bde817c221d9d8344e
Signed-off-by: Cyril Roelandt <cyril@redhat.com>
Signed-off-by: Tony Breeds <tony@bakeyournoodle.com>