[check-uc*] Create essentially a "no-op" dependency
This works around an unknown aspect of tox in that until you do
something (install packages, or run commands) tox doesn't create
.tox/{envdir}/log.  This isn't a big deal except
tox_install_sibling_packages[1] relies on that existing.

If it doesn't we end up with a module failure:

Change-Id: I3824235b49b77e84e693a8a37c9ff3a1027a3acd
---
2025-10-23 21:38:45.404140 | TASK [tox : Install any sibling python packages]
2025-10-23 21:38:46.846253 | ubuntu-jammy | MODULE FAILURE:
<snip>
2025-10-23 21:38:46.847689 | ubuntu-jammy | FileNotFoundError: [Errno 2] No such file or directory: '/home/zuul/src/opendev.org/openstack/requirements/.tox/py310-check-uc/log/py310-check-uc-siblings.txt'
---

Until that module is fixed add 'setuptools' as a dependency.  It will be
installed already as part of the virtualenv so is funtionally equivilent
to an empty set of deps.

This is a enhancement? to Iaa7f02ec9d1e734884b1329e2b177964f54d64df
Also updated upper-constraints.txt to ensure the check-uc* jobs run

[1] https://opendev.org/zuul/zuul-jobs/src/branch/master/roles/tox/library/tox_install_sibling_packages.py#L387

Change-Id: I7fd709d0cf0468c90593b5f63bb34d58e9c1e774
Signed-off-by: Tony Breeds <tony@bakeyournoodle.com>