Move parsing of setup.cfg, requirements.txt files
At least part of it anyway. Rather than waiting to split the lines and
strip the non-dependency information (comments etc) from the file, we do
it immediately on load.

We also add a whole lot of type hints to the modified code, to help us
reason about things a little easier.

Change-Id: Ide4574c53123ebe20d28211b80b9c079d01dd5e0
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>