Correct expected type for Requirements.process
In change Ide4574c53123ebe20d28211b80b9c079d01dd5e0, we changed the type
of openstack_requirements.project.Project.requirements from a dict of
strings (corresponding to filename mapped to file contents) to a dict of
lists of strings (corresponding to filename mapped to file lines).
However, we forgot to update one of the callers. Do this, adding the
hints that should have been added initially to catch this obvious issue.

Change-Id: I5f9aff4d964c049f7681126c74a2ee89f3ba35f1
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>