Don't include type stub packages in upper-constraints
We don't need to version these as they don't affect runtime. Heck, in
most cases they shouldn't be included in the requirements.txt files,
though we still need to denylist them here if we want to include them in
e.g. a typing extra.

We also remove 'typing' from global-requirements since this is a stdlib
package, not a third-party one [1]. typing-extensions is kept since this
is third-party as is used at runtime (though we correct the name used in
upper-constraints).

[1] https://docs.python.org/3/library/typing.html

Change-Id: I01dc6aa8e6389fc927d03c4997052d237639b9fe
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>